moved countdown file and starmer
