+++ src/components/kanaquiz-game/timer.tsx
@@ -1,0 +1,7 @@
+export default function Timer() {
+  return (
+    <>
+      <h1>0:00</h1>
+    </>
+  );
+}
